Added env variable to disable syncing from git

+++ sync/synchronize.sh
@@ -20,16 +20,18 @@ NOTES=${ROOT}/content/notes
 # Go to root repo directory for execute all commands
 cd ${ROOT}
 
-# Sync static generator
-git reset --hard
-git fetch --all
-git rebase
+if [[ -z "${NOSYNC}" ]]; then
+  # Sync static generator
+  git reset --hard
+  git fetch --all
+  git rebase
 
-# Renew blog posts and themes
-git submodule sync
-git submodule init
-git submodule update
-git submodule foreach "(git checkout master && git pull --ff origin master && git push origin master) || true"
+  # Renew blog posts and themes
+  git submodule sync
+  git submodule init
+  git submodule update
+  git submodule foreach "(git checkout master && git pull --ff origin master && git push origin master) || true"
+fi
 
 # Copy images to static
 cp -a ${BLOG}/images ${ROOT}/static
